# Optimized map::insert