```cpp
template<int N, class L = brigand::range<int, 0, N>>
struct mk_map;

template<int N, class... T>
struct mk_map<N, brigand::list<T...>> {
  using type = brigand::map<brigand::pair<T,T>...>;
};

template<class T> using fn = decltype(mk_map<SIZE>::type::insert(brigand::type_<brigand::pair<T,int>>{}));

using list = brigand::transform<mk_map<SIZE>::type, brigand::quote<fn>>;
```

Gcc:

Size = |      100      |      200      |      300
------ | ------------- | ------------- | -------------
before | 00.19s -  49M | 00.51s - 115M | 00.98s - 266M
after  | 00.12s -  38M | 00.28s -  78M | 00.53s - 167M

Clang:

Size = |      100      |      200      |      300
------ | ------------- | ------------- | -------------
before | 00.18s -  48M | 00.43s -  75M | 00.76s - 116M
after  | 00.12s -  41M | 00.22s -  51M | 00.35s -  64M